Implemented expression calculator

+++ test/variants/TypeSafeVariantOperations_test.go
@@ -8,16 +8,16 @@ import (
 )
 
 func TestSafeOperations(t *testing.T) {
-	a := variants.NewVariant(int32(123))
+	a := variants.NewVariant(123)
 	manager := variants.NewTypeSafeVariantOperations()
 
 	b := manager.Convert(a, variants.Float)
 	assert.Equal(t, variants.Float, b.Type())
 	assert.Equal(t, float32(123.0), b.AsFloat())
 
-	c := variants.NewVariant(int32(2))
-	assert.Equal(t, int32(125), manager.Add(a, c).AsInteger())
-	assert.Equal(t, int32(121), manager.Sub(a, c).AsInteger())
+	c := variants.NewVariant(2)
+	assert.Equal(t, 125, manager.Add(a, c).AsInteger())
+	assert.Equal(t, 121, manager.Sub(a, c).AsInteger())
 	assert.False(t, manager.Equal(a, c).AsBoolean())
 
 	array := []*variants.Variant{
