+++ test/variants/Variant_test.go
@@ -10,8 +10,8 @@ import (
 func TestVariants(t *testing.T) {
 	a := variants.NewVariant(123)
 	assert.Equal(t, variants.Integer, a.Type())
-	assert.Equal(t, int32(123), a.AsInteger())
-	assert.Equal(t, int32(123), a.AsObject())
+	assert.Equal(t, 123, a.AsInteger())
+	assert.Equal(t, 123, a.AsObject())
 
 	b := variants.NewVariant("xyz")
 	assert.Equal(t, variants.String, b.Type())
